fix: deduplicate code in available jobs
