+++ packages/apps/human-app/frontend/src/modules/worker/components/jobs/available-jobs/desktop/available-jobs-table.tsx
@@ -166,7 +166,7 @@ const getColumns = (
 
 export function AvailableJobsTable({
   chainIdsEnabled,
-}: AvailableJobsTableProps) {
+}: Readonly<AvailableJobsTableProps>) {
   const { colorPalette, isDarkMode } = useColorMode();
   const {
     setSearchEscrowAddress,
