+++ packages/apps/human-app/frontend/src/modules/worker/components/jobs/available-jobs/mobile/available-jobs-table-mobile.tsx
@@ -8,8 +8,6 @@ import { useJobsFilterStore } from '@/modules/worker/hooks/use-jobs-filter-store
 import { Alert } from '@/shared/components/ui/alert';
 import { getNetworkName } from '@/modules/smart-contracts/get-network-name';
 import { getErrorMessageForError } from '@/shared/errors';
-import type { AvailableJob } from '@/modules/worker/services/fetch-available-jobs';
-import { useInfiniteAvailableJobsQuery } from '@/modules/worker/hooks/use-infinite-available-jobs-query';
 import { Loader } from '@/shared/components/ui/loader';
 import { EvmAddress } from '@/modules/worker/components/jobs/evm-address';
 import { Chip } from '@/shared/components/ui/chip';
@@ -19,6 +17,10 @@ import { useColorMode } from '@/shared/hooks/use-color-mode';
 import type { JobType } from '@/modules/smart-contracts/EthKVStore/config';
 import { EscrowAddressSearchForm } from '@/modules/worker/components/jobs/escrow-address-search-form';
 import { AvailableJobsAssignJobButton } from '@/modules/worker/components/jobs/available-jobs/mobile/available-jobs-assign-job-button';
+import {
+  type AvailableJob,
+  useInfiniteAvailableJobsQuery,
+} from '@/modules/worker/services/available-jobs-data';
 
 interface AvailableJobsTableMobileProps {
   setIsMobileFilterDrawerOpen: Dispatch<SetStateAction<boolean>>;
