--- packages/apps/human-app/frontend/src/modules/worker/services/fetch-available-jobs/index.ts
@@ -1,3 +1,0 @@
-export * from './fetch-available-jobs.types';
-export * from './fetch-available-jobs.schema';
-export * from './fetch-available-jobs';
